requirements.txt
azure-cognitiveservices-speech==1.40.0
keyboard==0.13.5
numpy==2.0.0
pywin32==306
sounddevice==0.5.0